+++ tests/tests.rs
@@ -1,4 +1,4 @@
 #[test]
 fn can_set_as_hook() {
-    web_panic_report::set_hook_with("", |_| {});
+    web_panic_report::set_hook_with("", "", |_| {});
 }